refactor: optimize forecast report queries
